students/k3340/Meshcheryakov_Daniil/Lr3/task_2/requirements.txt
Django==4.2.7
djangorestframework==4.14.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
django-filter==23.5
Pillow==10.1.0

